Add files via upload
This is a search query that will get the links from google.com. 
Link it to the search bar file I uploaded earlier and pass the query's into this file and you will get all the links from google pertaining to the term searched. Then you can do whatever you want with that